requirements.txt
Flask
ipwhois
geoip2
requests
dnspython
pyOpenSSL
Flask-Login
Flask-WTF
email_validator
